Move actions to Policy file
